requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   nltk
    #   uvicorn
defusedxml==0.7.1
    # via youtube-transcript-api
filetype==1.2.0
    # via langchain-google-genai
google-ai-generativelanguage==0.9.0
    # via langchain-google-genai
google-api-core==2.28.1
    # via google-ai-generativelanguage
google-auth==2.42.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
googleapis-common-protos==1.71.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.76.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   grpcio-status
grpcio-status==1.76.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langsmith
    #   mcp
httpx-sse==0.4.3
    # via mcp
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
joblib==1.5.2
    # via nltk
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.25.1
    # via mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
langchain-core==1.0.1
    # via
    #   youtube-summary-mcp (pyproject.toml)
    #   langchain-google-genai
langchain-google-genai==3.0.0
    # via youtube-summary-mcp (pyproject.toml)
langsmith==0.4.38
    # via langchain-core
mcp==1.19.0
    # via youtube-summary-mcp (pyproject.toml)
nltk==3.7
    # via youtube-summary-mcp (pyproject.toml)
orjson==3.11.4
    # via langsmith
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==6.33.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.3
    # via
    #   youtube-summary-mcp (pyproject.toml)
    #   langchain-core
    #   langchain-google-genai
    #   langsmith
    #   mcp
    #   pydantic-settings
pydantic-core==2.41.4
    # via pydantic
pydantic-settings==2.11.0
    # via
    #   youtube-summary-mcp (pyproject.toml)
    #   mcp
python-dotenv==1.2.1
    # via pydantic-settings
python-multipart==0.0.20
    # via mcp
pyyaml==6.0.3
    # via langchain-core
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.10.23
    # via nltk
requests==2.32.5
    # via
    #   google-api-core
    #   langsmith
    #   requests-toolbelt
    #   youtube-transcript-api
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.28.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
sse-starlette==3.0.2
    # via mcp
starlette==0.49.1
    # via mcp
tenacity==9.1.2
    # via langchain-core
tqdm==4.67.1
    # via nltk
typing-extensions==4.15.0
    # via
    #   grpcio
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via mcp
youtube-transcript-api==1.2.3
    # via youtube-summary-mcp (pyproject.toml)
zstandard==0.25.0
    # via langsmith
